refactor(core/templates): extract common templates from profiles
