+++ src/core/templates/show-link.js
@@ -2,7 +2,7 @@
 import { html } from "../../core/import-maps.js";
 import { pub } from "../../core/pubsubhub.js";
 
-export default link => {
+export default function showLink(link) {
   if (!link.key) {
     const msg =
       "Found a link without `key` attribute in the configuration. See dev console.";
@@ -14,12 +14,12 @@ export default link => {
     <dt class="${link.class ? link.class : null}">${link.key}:</dt>
     ${link.data ? link.data.map(showLinkData) : showLinkData(link)}
   `;
-};
+}
 
 function showLinkData(data) {
   return html`<dd class="${data.class ? data.class : null}">
     ${data.href
       ? html`<a href="${data.href}">${data.value || data.href}</a>`
-      : ""}
+      : data.value}
   </dd>`;
 }
