+++ src/w3c/templates/cgbg-headers.js
@@ -1,9 +1,9 @@
 // @ts-check
 import { html } from "../../core/import-maps.js";
 import { l10n } from "./headers.js";
-import showLink from "./show-link.js";
-import showLogo from "./show-logo.js";
-import showPeople from "./show-people.js";
+import showLink from "../../core/templates/show-link.js";
+import showLogo from "../../core/templates/show-logo.js";
+import showPeople from "../../core/templates/show-people.js";
 
 export default (conf, options) => {
   const existingCopyright = document.querySelector(".copyright");
