+++ src/w3c/templates/headers.js
@@ -2,9 +2,9 @@
 import { getIntlData, humanDate } from "../../core/utils.js";
 import { html } from "../../core/import-maps.js";
 import { pub } from "../../core/pubsubhub.js";
-import showLink from "./show-link.js";
-import showLogo from "./show-logo.js";
-import showPeople from "./show-people.js";
+import showLink from "../../core/templates/show-link.js";
+import showLogo from "../../core/templates/show-logo.js";
+import showPeople from "../../core/templates/show-people.js";
 
 const ccLicense = "https://creativecommons.org/licenses/by/4.0/";
 const w3cLicense = "https://www.w3.org/Consortium/Legal/copyright-documents";
